feat(participants): setup model associations
Adds a new field to `Participant` called `Membership` which will be used
in the feature to enable GORM associations on queries. Right now the
queries that would use this asscoiation are custom SQL queries that
can't be changed without introducing breaking changes. We will circle
back to this in the future.

Closes #351